Fallout3 havok materials
Added complete list of havok materials for Fallout3. Material definitions are stored in bits 0-4, bit 5 is flag "PLATFORM" and bit 6 is flag "STAIRS". Nifskope unfortunately is not able to edit some bit range as dropdown menu and other bits as "tickable" selection. So, in Fallout3HavokMaterial enum there is defined all 127 combinations.