Modify of Num UV Sets
Changed definition of "Num UV Sets" (for >= 10.0.1.0 only) and get rid of "BS Num UV Sets".
"Num UV Sets" defined as byte containing number of UV texture sets. Added definition of "Extra Vectors Flags" stored in byte. Only known function of that flags by that time is 4th bit representing flag for Tangents/Bitangents arrays - this is defined in bitflags type "ExtraVectorsFlags".
As there is no certainity that number of UV Sets can be stored in all 8 bits, I have left AND function on line 2731 where "arr1" is getting a number of UV Sets only from lower 6 bits (i.e. max number of UV Sets is 63). There is a note about that on line 2725 (although it is for old version 4.0.0.2).