[FO4] New shader flags
There were 16 changed shader flags between game versions.  This creates
a new set of enums for FO4 so that the proper descriptions can be shown
for the bits.

Also, the original vercond for Shader Flags 1/2 was unnecessary.  The
Skyrim NIFs that are User Version == 11 are actually structured like FO3
NIFs and use different shader blocks.  So changing the condition like I
did is a sound change, and I've already scanned all vanilla Skyrim NIFs.