Removal of nifskopetype
In <basic> and <compound> the nifskopetype attribute was used as an
alias system to remap types to internal types.   NifSkope now handles
the remapping internally.   One single downside to this is that new
basic/compound cannot be remapped to an internal type via XML alone, it
must be added to NifSkope.  However, if this proves to ever be an issue,
NifSkope can house another file in the installation where the remapping
can be done.

Additionally, nifskopetype was used in some <add> rows to denote that
the data was a binary blob.  I have instead replaced this with a more
generic `binary="1"`.  This is also more informative as generally these
rows are actually embedded files.

Lastly I removed a lot of custom "ns" compounds which NifSkope no longer
seems to use.  It would seem these compounds were even pre-nifskopetype
attribute.