[SSE] Initial Commit
Support for SSE.  The BSVertexData had to be moved up for the
SkinPartition.  The changes have been tested on all SSE files and used
for the past several months with no issue.  There are no known FO4
regressions.

A new token for division in expressions was introduced, `#DIV#`, as "/"
is used in some XML attributes already and also causes parsing issues
with NifSkope which already uses that token for things at runtime.

The BSSegment PR from ttl269 has also been included in this as it is a
valid change and required for SSE.