Num UV Sets fix
This is a redo of #51 since I did not seem to have permission to rebase the PR.   These changes work for all games and fix all FO3 UV Sets regressions.

The only downside is for UV Sets > 1 in non-Bethesda games the UV Sets count now becomes a combination of flags.  UV_1 + UV_2 = 3 UV Sets,  and UV_1 + UV_4 = 5 UV Sets and so on.

This was the only way to treat the field as a single ushort which is necessary for endianness changes when reading the field 16 bytes at once because the value is read/written 16 bytes at once and uses bitmasking to retrieve the UV count.