Fix Fallout3Layer
Removed BIPED_PART enum values from end of Fallout3Layer.  These are actually a different enum for the next byte (Flags and Part Number).  Max layers is 64 so we know this to be a mistake.  There were also no usages of Fallout3Layer in FO3/NV above FOL_NULL (43).

Renamed value 29 to be more accurate.