Relaxed suffix enforcement for niflib
Duplicate name avoidance was too strict and unnecessary in these cases.  Nifdocsys will now attempt to cast subsequent duplicate names to the correct types in order to ensure correct read/write size.  As such the first type should always be the largest, so the "Flags" fields on NiAVObject were switched.

So in the case of Flags and Body Flags, a uint data member will be created on the class and will be read/written to with a cast during I/O.

The suffix is still required when the types are not castable between one another.