Header compound overhaul
Almost completely redone with correct version ranges, fixed names, fixed types, and corrections for 30.x NIFs.   30.x NIFs with embedded thumbnail metadata would not load before these fixes.

Also, BSStreamHeader was annotated but a compound for it was not feasible for several reasons, namely the nesting of `User Version 2` which would break existing parsers.   This also leads to the ugly repetition of the cond expression for it.