NiBSpline* type, name, and doc fixes
Offset and Half Range are the correct names for the compact control point data used to decompress the compact control points.

Beware for NiBSplineTransformInterpolator and NiBSplineCompTransformInterpolator the naming has shifted for "Offset".

"Offset" is now "Handle" and "Bias" is now "Offset".   This means the same names still exist but for different rows.

Also shifted some types around to the correct classes and filled out some unknowns.