Use prefix for some Bethesda enums
With gen_niflib supporting the prefix attribute, enums can be shifted to using it where needed.  In this case the name attributes of the bitflags take up a lot of space in the UI and removing the prefix compacts this.

Also made them more consistent.  FO3 is now  F3SF1/F3SF2 and FO4 is F4SF1/F4SF2.