Add proper defaults, initial pass
Aids in proper creation of new blocks by having sound default values.

Also splits Glossiness/Smoothness in BSLightingShaderProperty to have correct defaults and documentation on ranges.  Unknown Float 1 was changed to Rimlight Power as well, and the next line, Backlight Power, relies on its value being FLT_MAX.  So  `Rimlight Power == 0x7F7FFFFF` is used as a `cond` for Backlight Power and this may require the parser to deal with a float to hex comparison.