NiEvaluator and 20.5+ KF support, better animation defaults.
Added NiSequenceData and NiEvaluator for 20.5.0.0+.  In 20.5.0.0 NiSequenceData is an alias for NiControllerSequence.  The XML does not handle this and will defer to parser support.  It does not seem to happen in practice.

For 20.5.0.1,  NiSequenceData is a hybrid of old and new,  pre-NiEvaluator.

For 20.5.0.2+, NiEvaluator replaces NiInterpolator for everything.  Data wise they are almost exactly the same.

Also improved the default values for Blend and BSpline interpolators.