[PARSER CHANGES REQUIRED] Use `\` as compound member accessor
Instead of requiring all parsers to special-case the arg passing for BSVertexDesc,  have just adopted the `\` syntax for compound member accessor.

So at least for parsers that do not internalize BSVertexDesc to be `uint64_t` or equivalent to mirror the game,  all you need to do is support `\`.

For C++ this would go from

`Vertex Desc\Vertex Attributes`
to
`vertexDesc.vertexAttributes`

The period syntax was decided against because it is not very visible and `\` is familiar from filepaths / directory trees.