NiPhysXScene, NiPhysXSceneDesc, NiPhysX particles
Scene/SceneDesc were relatively old and unused and also quite complex to implement so they never got ported to XML from the 010 templates.

Note:  Half the files that actually use Scene/SceneDesc are in Empire Earth III, which its version 20.2.0.8 behaves differently for these blocks than other 20.2.0.8 versions.  The only way to tell the difference is by the file extension, which is .nifcache.   This is beyond the scope of nif.xml, and so only the game files that match the version conditions of Gamebryo proper are supported.