Fixes for 20.3.0.9 custom versions
Only MdlManBinary.nif cannot be read for Divinity 2, and it appears to be some kind of rig filled with custom classes serialized to NIF.  There is no desire to decode this single NIF with 10 custom classes.

Also the Divinity 2 XML files that have a Gamebryo header are ignored as they appear to just contain XML strings in more custom classes.