Initial implementation for #69
Made `<version>` mean distinct combination of header values which produce identical block layouts.  This meant splitting up several main Versions into many based on User Version or Bethesda  Version.

id = Used to uniquely identify this collection of NIF versions. Can also be used for enum generation, map/dictionary keys.

user = List of User Versions which are functionally equivalent

bsver = List of Bethesda Versions which are functionally equivalent

custom = If user/bsver are default but the Version is actually specific to a developer and not an official Gamebryo version

supported = Whether or not the XML fully supports reading/writing of this version.

ext = List of custom file extensions associated with this version.