Simplifying BSVertexData conds
They were kept separate for clarity in actual code using the enumerations; clarity that doesn't transfer over to XML which is using the raw integers.

```
(A & X) != 0 && (A & Y) != 0 && (A & Z) == 0
```
Is the same as
```
(A & (X|Y|Z)) == (X|Y)
```
Basically != 0 you want that particular bit on the RHS of the expression, and == 0 you do not.
Anything tested you put on the LHS.

Also switched the order of HalfVector3/Vector3 as the larger type should come first.