Primary version indicators for #69
A concise way of listing which versions are most important to each game.

Each game should have only one set of `{{}}`.

Note:  In the case of Oblivion, there are two "primary" versions..  20.0.0.4 (10) is used for KF and 20.0.0.5 is used for NIF.  Thus I changed the "game name" to `{{Oblivion KF}}` for 20.0.0.4 (10).