[Linting] Inherited duplicates
Fixed inherited duplicate names, as well as the layout of BSTriShape for Particle emitter geometry data in SSE.

Left the name for Dynamic Data as "Vertices" though it is technically just called "Dynamic Data" in engine.  It contains the vertices and bitangent X for any BSDynamicTriShape, so the naming is apt.