Move and Rename User Version 2, change ExportInfo to BSStreamHeader
Now that any usage of `User Version 2` was actually just a `#BSVER#` token, reorganized the header information for Bethesda NIFs into the correct layout, and renamed the member to a more appropriate BS Version.

The global for #BSVER# uses the member of operator `\`, but as it is a token now it is free to be used however one wishes.

I split #BSSTREAMHEADER# into its own token in case the expression is ever reused (it used to be on multiple members before this reorganization).

Also put `Max Filepath` into BSStreamHeader where it belonged.