Add `convertible`, split `countable`
For #76

The signed integeral types were marked as countable even though they should not be used for array sizes.  A negative value cannot be passed into an array size.   They are still integral though, so countable was split into `countable` and `integral`.

In addition,  added `boolean` to designate types that can toggle the presence of another member.

Also reordered items for `convertible` so that the referenced types come before the type referencing it, i.e. larger types first.