Improved NiAdditionalGeometryData
Named the compounds correctly after the actual structs,  separated the "Has Data" bool from the compound as it's not actually part of the struct, but before it in a for loop during (de)serialize.

Combined NiAdditionalGeometryData and the BS version into one compound when possible.  But the block can actually no longer be found in any official files, so the BS version is not very important.