BSLightingShader decoding
FO4 DLC uses BSLightingShaderPropertyUShortController in 2 NIFs, but according to the CK it is not actually a functional block and spits out constant warnings. It is still needed in nif.xml for serialization though.

FO4 also adds two booleans for Env Map shader type for SSR.  These were previously marked as an Unknown Short.  These booleans are *not* synced with their values in the BGSM files except for generated meshes such as SCOL and precombined.