Replace `calculated` with `calc`
For #76.  `calc` does not need to be supported and is only used for pre-serialization preparation of the data.  This can also be done by hand.

Some revisions also need to be made for #70 and #73 to account for another attribute with its own expression grammar and tokens.

However, the tokens introduced for `calc` will not be added to the tokens in the XML and will need to be explicitly supported if supporting `calc`.  They are:

1D array size (uses regex): `#LEN[(.*?)]#`
2D array size (uses regex): `#LEN2[(.*?)]#`

Ternary `?`: `#THEN#`
Ternary `:`: `#ELSE#`

Also removes any unnecessary `calculated` without replacement.

Additionally, `calc` is used to limit array size for the time being, though it's possible this should be done with its own attribute, such as `maxlen`.