Default Values Pass 1
Added `default` to a large number of members, as well as `calc` to a few members to enforce array sizes based on other data.

Also fixed a few `type` issues.

This is a first pass in a much larger effort,  using data analysis to gather the most commonly used values.