BSVertexDesc correct bitfield
Made a correct bitfield using `uint64` as the underlying type, split into nibbles for the sizes and offsets of the vertex attributes.

Changed VertexFlags to VertexAttributes and it no longer has the empty nibble in front.

ARG passing of the BSVertexDesc is treated as the underlying numeric value of the bitfield, so that the attributes can be passed into BSVertexData using `#RSH# 44` instead of the member accessor syntax.

This makes the first use of the basic `uint64` type, so (de)serialization now requires support of 64-bit values.