String type fixes
Specify SizedString vs NiFixedString where appropriate.  The compound type `string` is an extra level of indirection in many cases, where the block can never be used on versions less than 20.1.0.3 where the CString to NiFixedString transition happened.

NiFixedString is a basic type, so using it where possible is recommended over the `string` compound type.

`string` could NOT be changed when dealing with multiple rows with the same name but version exclusive, such as "Accum Root Name".

Changed ShortString to ExportString as it was already documented as being exclusive to the Bethesda stream header.

Added SizedString with ushort length type, for the SSF File path.