Legacy NiParticle decoding
NetImmerse NiParticle (called NiOldParticle after Gamebryo 1.1) was severely undecoded.  This affects several NetImmerse games like Morrowind and Freedom Force.

The per-particle info in NiOldParticle actually remained identical to the class in NiParticle, which is named "NiParticleInfo".  It did however have a different name in NetImmerse which was "NiPerParticleData".  Since they were the same data, the two compounds were merged and NiPSysData was updated.