Misc decoding, custom blocks for WorldShift, Guild 2
Decoded some unknown fields and blocks for WorldShift and Guild 2.

NiSourceTexture is now correctly decoded.  The unknown byte actually controls whether the image data Ref gets serialized or not, but it was always 1.  The addition of another conditional value increases the complexity for Pixel Data to 3 rows, all exclusive by version or condition.